feat(core): add transport layer and refactor frame sending
- Add transport package with related classes and interfaces
- Implement NodesManager and TransportProvider in MeshCore
- Update FrameRouter to use TransportProvider for sending frames
- Modify CoreEvents and CoreInput to use bytes instead of raw frames
- Refactor advertising functionality in CoreEvents

+++ core/src/main/java/io/github/teilabs/meshnet/core/frame/FrameConstants.java
@@ -2,7 +2,7 @@ package io.github.teilabs.meshnet.core.frame;
 
 /** Class with all Frame related constants. */
 public final class FrameConstants {
-    public static final int VERSION = 1;
+    public static final byte VERSION = 1;
 
     // Version 1
     // Sizes
